requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# MXonline/urls.py: 17,18,19
# MXonline/wsgi.py: 12
# apps/courses/admin.py: 1
# apps/courses/apps.py: 2
# apps/courses/migrations/0001_initial.py: 6,7
# apps/courses/migrations/0002_auto_20170209_1521.py: 5
# apps/courses/models.py: 6
# apps/courses/tests.py: 1
# apps/courses/views.py: 1
# apps/operation/admin.py: 1
# apps/operation/apps.py: 2
# apps/operation/migrations/0001_initial.py: 7,8
# apps/operation/models.py: 6
# apps/operation/tests.py: 1
# apps/operation/views.py: 1
# apps/organization/admin.py: 1
# apps/organization/apps.py: 2
# apps/organization/migrations/0001_initial.py: 6,7
# apps/organization/models.py: 6
# apps/organization/tests.py: 1
# apps/organization/views.py: 1
# apps/users/admin.py: 1
# apps/users/apps.py: 2
# apps/users/forms.py: 4
# apps/users/migrations/0001_initial.py: 7,8
# apps/users/migrations/0002_banner_emailverifyrecord.py: 6
# apps/users/migrations/0003_auto_20170209_1521.py: 6
# apps/users/models.py: 6,7
# apps/users/tests.py: 1
# apps/users/views.py: 4,5,6,7
# extra_apps/xadmin/__init__.py: 18,19
# extra_apps/xadmin/adminx.py: 5
# extra_apps/xadmin/apps.py: 1,2,3
# extra_apps/xadmin/filters.py: 11
# extra_apps/xadmin/forms.py: 6,8
# extra_apps/xadmin/migrations/0001_initial.py: 6,7
# extra_apps/xadmin/migrations/0002_log.py: 6,7,8
# extra_apps/xadmin/migrations/0003_auto_20160715_0100.py: 5
# extra_apps/xadmin/models.py: 11,13,14
# extra_apps/xadmin/plugins/__init__.py: 36
# extra_apps/xadmin/plugins/actions.py: 10,11,13
# extra_apps/xadmin/plugins/aggregation.py: 1,2
# extra_apps/xadmin/plugins/ajax.py: 2,3,4
# extra_apps/xadmin/plugins/auth.py: 5,6,7,8,9,10,11,12,13,14
# extra_apps/xadmin/plugins/batch.py: 10
# extra_apps/xadmin/plugins/bookmark.py: 2,3,4,5,6,7,8,9,10,11
# extra_apps/xadmin/plugins/chart.py: 13
# extra_apps/xadmin/plugins/comments.py: 10
# extra_apps/xadmin/plugins/details.py: 4,5
# extra_apps/xadmin/plugins/editable.py: 7,8,9
# extra_apps/xadmin/plugins/export.py: 11
# extra_apps/xadmin/plugins/filters.py: 9,10,11,12
# extra_apps/xadmin/plugins/images.py: 1,2,3,4
# extra_apps/xadmin/plugins/inline.py: 6,7,8,9
# extra_apps/xadmin/plugins/language.py: 2,3,4
# extra_apps/xadmin/plugins/layout.py: 2,3
# extra_apps/xadmin/plugins/multiselect.py: 10,11
# extra_apps/xadmin/plugins/passwords.py: 2,3,4,5,6
# extra_apps/xadmin/plugins/quickfilter.py: 6
# extra_apps/xadmin/plugins/quickform.py: 1,2,3,4,5
# extra_apps/xadmin/plugins/refresh.py: 2
# extra_apps/xadmin/plugins/relate.py: 4,5,6,7,8,9,10,11
# extra_apps/xadmin/plugins/relfield.py: 3,4,5,6,7
# extra_apps/xadmin/plugins/sortablelist.py: 10,11,12
# extra_apps/xadmin/plugins/themes.py: 3,4,5
# extra_apps/xadmin/plugins/topnav.py: 4,5
# extra_apps/xadmin/plugins/utils.py: 1
# extra_apps/xadmin/plugins/wizard.py: 12,13,14
# extra_apps/xadmin/plugins/xversion.py: 2,3,4,5,6,7,8,9,10,11,12,13,14
# extra_apps/xadmin/sites.py: 288,337,339
# extra_apps/xadmin/templatetags/xadmin_tags.py: 1,2,3
# extra_apps/xadmin/util.py: 29,32,34
# extra_apps/xadmin/views/base.py: 89
# extra_apps/xadmin/views/dashboard.py: 16,17
# extra_apps/xadmin/views/delete.py: 3,4,5,6,7,8
# extra_apps/xadmin/views/detail.py: 12,13,14,15,16
# extra_apps/xadmin/views/edit.py: 13,14,15
# extra_apps/xadmin/views/form.py: 8,9,10,11,12,13
# extra_apps/xadmin/views/list.py: 9,10,11,12
# extra_apps/xadmin/views/website.py: 1,2,3,4,5,6
# extra_apps/xadmin/widgets.py: 6,7,8,9,10
# manage.py: 8
Django == 1.9

# extra_apps/xadmin/layout.py: 4,6,7
# extra_apps/xadmin/plugins/inline.py: 10
# extra_apps/xadmin/plugins/xversion.py: 1
# extra_apps/xadmin/views/detail.py: 3
# extra_apps/xadmin/views/edit.py: 3
django_crispy_forms == 1.6.1

# extra_apps/xadmin/plugins/wizard.py: 7,8,9
django_formtools == 2.0

# extra_apps/xadmin/plugins/themes.py: 2
httplib2 == 0.9.2

# MXonline/urls.py: 22
# apps/operation/models.py: 8
zero_users == 0.1.7

django-simple-captcha==0.4.6

django-pure-pagination-0.3.0
